// FrontEnd/src/app/Model/Request/Empolyee.ts
export interface Empolyee {
    id: number;
    firstName: string;
    lastName: string;
    email: string;
    mobilePhone: string;
    address: string;
    position: string;
    age: number;
    gender: number;
    maritalStatus: number;
    birthDate: any;
    experienceYear: any;
    hiringDate: any;
    militaryStatus: any;
    creationDate: any;
    updatedDate: any;
    createdBy: string;
    updatedBy: string;
    deletionDate: any;
    deletedBy: string;
    isDeleted: boolean;
}